Switch to poetry and add tests
This commit transitions to use poetry for building and dependency management. https://python-poetry.org

It also adds a rudimentary integration test that emulates a Modbus meter over TCP and exercises a meter
module to parse the registers mocked there and output a valid result.
