--- setup.cfg
@@ -1,31 +1,0 @@
-[metadata]
-name = sdm_modbus
-version = 0.4.6
-author = nmakel
-description = Eastron SDM Modbus parser library
-long_description = file: README.md
-long_description_content_type = text/markdown
-license = MIT License
-url = https://github.com/nmakel/sdm_modbus
-project_urls =
-    Issue Tracker = https://github.com/nmakel/sdm_modbus/issues
-classifiers =
-    Development Status :: 4 - Beta
-    Programming Language :: Python :: 3.7
-    Operating System :: OS Independent
-    Environment :: Console
-    Natural Language :: English
-    Intended Audience :: Developers
-    License :: OSI Approved :: MIT License
-
-[options]
-package_dir =
-    = src
-packages = find:
-include_package_data = True
-python_requires = >= 3.7
-install_requires =
-    pymodbus >= 2.4.0
-
-[options.packages.find]
-where = src
