+++ tox.ini
@@ -1,0 +1,9 @@
+[tox]
+skipsdist = true
+envlist = py37
+
+[testenv]
+whitelist_externals = poetry
+commands =
+    poetry install -v
+    poetry run behave tests/features
